requirements.txt
requests
PyYAML
pymongo
matplotlib
